requirements.txt
atlassian-python-api
openai
python-dotenv
nltk
transformers
numpy
flask
pandas
pandas
datetime
